feat: add cart types, Razorpay utility, and order summary components
